VigilCD/pyproject.toml
[project]
name = "vigilcd"
version = "0.7.0"
description = "VigilCD - GitOps Deployment Agent"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = "==3.14.*"
authors = [
    { name = "Ruben Sprengel" },
]
dependencies = [
    "apscheduler==3.11.1",
    "fastapi[standard]==0.124.4",
    "gitpython==3.1.45",
    "pydantic==2.12.5",
    "pyyaml==6.0.3",
    "python-dotenv==1.2.1",
    "structlog==25.5.0",
    "uvicorn[standard]==0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "ruff==0.14.9",
    "mypy==1.19.1",
    "types-PyYAML==6.0.12.20250915"
]

vault = [
    "hvac==2.4.0",  # HashiCorp Vault Client
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src"]

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 100
indent-width = 4
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]


[tool.ruff.lint]
select = [
    "D",
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
    "SIM",
    "S",
    "PL",
    "FAST"
]

ignore = [
    "E501",
    "D203",
    "D212",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004", "S105", "S106", "PLR0913"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-local-folder = ["src"]

[tool.ruff.lint.pylint]
allow-dunder-method-names = ["__tablename__", "__table_args__"]
